I thought about implementing an algorithm to guarantee that all words are placed. There are problems doing this, but it would look something like this:
1) Insert word at random location/orientation
2) Attempt to insert next word at random location/
3) If step #2 fails, repeat step #1
4) Continue to place items

The issue is that by placing randomly, there is no guarantee how long
it will take to visit all possible locations. In addition, if a word is
added which forces the algorithm to pop all the way back to the first
word; this would be highly inefficient.

The next idea was to force a complete search. So modify the above
algorithm as follows:
1) Insert word at random location/orientation
2) Attempt to insert next word at random location/
3) If step #2 fails:
a) Move word from step 1:
   record the initial x, y position and iterate over the entire grid
   starting from that initial x,y. That way, each time we get popped off
   we start at the next position in the grid. If we eventually reach
   the start x,y - then the word that poppe us cannot be placed in the
   grid.

This could take a considerable amount of time and likely would mean that
the grid would need to be duplicated to revert if necessary.

In the end I tried several concepts - front loading the longer words,
flipping the direction of the word if it doesn't fit on first try, etc.
These concepts did help, although did very little for increasing the
number of words inserted.

Finally, I took a look at the % utilized in the grid and found that with
my very large word set it was 100% utilized - meaning that EVERY letter
in the grid was used. This led me to realize that the insert randomly
method was actually the fastest and gained the same insertion rate of
a full on iterative search within -10%. That's pretty darn good! So,
in the end, I'm going with the original approach.